refactor: useLetterDraft, useFieldIntro, KeyboardAccessoryBar 파일 생성하여 코드 역할 분리 및 글 본문 30자 체크
