+++ src/features/post/components/write/KeyboardAccessoryBar.module.css
@@ -1,0 +1,53 @@
+/* 모바일 키보드 바로 위 액세서리 바 */
+.bar {
+  flex-shrink: 0;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 0.8rem;
+  padding: 0.8rem 1.2rem;
+  background: var(--background-color);
+  border-top: 1px solid var(--color-primary-200);
+}
+
+.camera {
+  width: 4.4rem;
+  height: 3.6rem;
+  padding: 0;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  background: var(--color-primary-100);
+  border: 1px solid var(--color-primary-200);
+  border-radius: 0.8rem;
+  color: var(--color-primary-500);
+  cursor: pointer;
+}
+
+.nav {
+  display: flex;
+  gap: 0.8rem;
+}
+
+.btn {
+  width: 4.4rem;
+  height: 3.6rem;
+  padding: 0;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  background: var(--color-primary-400);
+  border: none;
+  border-radius: 0.8rem;
+  cursor: pointer;
+}
+.btn:disabled {
+  opacity: 0.35;
+  cursor: default;
+}
+.icon {
+  width: 2rem;
+  height: auto;
+  display: block;
+}
